Fix the window limits calculation for GTK w/ CSD
When CSD is enabled (which is the default for GTK on Wayland), the
titlebar height and borders are now included as part of the window,
whereas the caller would have intended for the given limits to only
cover the drawable area.

To fix this, we can find difference in sizes between the window and
child widget, and that size is used to offset the limits.

Fixes #84.